pyproject.toml
[project]
name = "alicelm"
version = "0.1.0"
description = "AI-powered video learning assistant for Bilibili"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AliceLM Team"}
]

dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    
    # HTTP Client
    "httpx>=0.25.0",
    "requests>=2.31.0",
    
    # Task Queue
    "apscheduler>=3.10.0",
    
    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    
    # Logging
    "structlog>=23.2.0",
    
    # Video Processing
    "you-get>=0.4.1650",
    "moviepy>=1.0.3",
    "pydub>=0.25.1",
    
    # ASR
    "openai-whisper>=20231117",
    "faster-whisper>=0.9.0",
    
    # AI/LLM
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Utils
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
